requirements-vercel.txt
# Minimal dependencies for Vercel HTTP API deployment
# This file is used by vercel.json for serverless function deployment

fastapi>=0.104.0
pydantic>=2.0.0
openap>=1.3.0
airportsdata>=20231201
geographiclib>=2.0
python-dotenv>=1.0.0

# Note: uvicorn is not needed for Vercel serverless functions
# Note: fastmcp, litellm, uvloop are excluded (not needed for HTTP API)
